Add support for filtering names using regexps. Add colors.

+++ internal/civo/kubernetes.go
@@ -1,84 +1,108 @@
 package civo
 
 import (
+	"errors"
 	"fmt"
 
+	"github.com/civo/civogo"
 	"github.com/konstructio/dropkick/internal/outputwriter"
 )
 
 // NukeKubernetesClusters deletes all Kubernetes clusters associated with the Civo client.
 // It returns an error if the deletion process encounters any issues.
 func (c *Civo) NukeKubernetesClusters() error {
-	c.logger.Printf("listing Kubernetes clusters")
+	c.logger.Infof("listing Kubernetes clusters")
 
 	clusters, err := c.client.ListKubernetesClusters()
 	if err != nil {
 		return fmt.Errorf("unable to list Kubernetes clusters: %w", err)
 	}
 
-	c.logger.Printf("found %d clusters", len(clusters.Items))
+	c.logger.Infof("found %d clusters", len(clusters.Items))
 
 	for _, cluster := range clusters.Items {
-		c.logger.Printf("found cluster: name: %q - ID: %q", cluster.Name, cluster.ID)
+		c.logger.Infof("found cluster: name: %q - ID: %q", cluster.Name, cluster.ID)
+
+		if c.nameFilter != nil && !c.nameFilter.MatchString(cluster.Name) {
+			c.logger.Warnf("skipping cluster %q: name does not match filter", cluster.Name)
+			continue
+		}
 
 		clusterVolumes, err := c.client.ListVolumesForCluster(cluster.ID)
 		if err != nil {
-			return fmt.Errorf("unable to list volumes for cluster %q: %w", cluster.ID, err)
+			return fmt.Errorf("unable to list volumes for cluster %q: %w", cluster.Name, err)
 		}
 
 		for _, volume := range clusterVolumes {
+			c.logger.Infof("found volume: name: %q - ID: %q", volume.Name, volume.ID)
+
+			if c.nameFilter != nil && !c.nameFilter.MatchString(volume.Name) {
+				c.logger.Warnf("skipping volume %q: name does not match filter", volume.Name)
+				continue
+			}
+
 			if c.nuke {
-				c.logger.Printf("deleting volume %q for cluster %q", volume.ID, cluster.ID)
+				c.logger.Infof("deleting volume %q for cluster %q", volume.Name, cluster.Name)
 				res, err := c.client.DeleteVolume(volume.ID)
 				if err != nil {
-					return fmt.Errorf("unable to delete cluster %q volume %q: %w", cluster.ID, volume.ID, err)
+					return fmt.Errorf("unable to delete cluster %q volume %q: %w", cluster.Name, volume.Name, err)
 				}
 
 				if res.ErrorCode != "200" {
-					return fmt.Errorf("Civo returned an error code %s when deleting volume %q: %s", res.ErrorCode, volume.ID, res.ErrorDetails)
+					return fmt.Errorf("Civo returned an error code %q when deleting volume %q: %s", res.ErrorCode, volume.Name, res.ErrorDetails)
 				}
 
-				outputwriter.WriteStdoutf("deleted volume %q for cluster %q", volume.ID, cluster.ID)
+				outputwriter.WriteStdoutf("deleted volume %q for cluster %q", volume.Name, cluster.Name)
 			} else {
-				c.logger.Printf("refusing to delete volume %q for cluster %q: nuke is not enabled", volume.ID, cluster.ID)
+				c.logger.Warnf("refusing to delete volume %q for cluster %q: nuke is not enabled", volume.Name, cluster.Name)
 			}
 		}
 
 		if c.nuke {
-			c.logger.Printf("deleting cluster %q", cluster.ID)
+			c.logger.Infof("deleting cluster %q", cluster.Name)
 			res, err := c.client.DeleteKubernetesCluster(cluster.ID)
 			if err != nil {
-				return fmt.Errorf("unable to delete cluster %q: %w", cluster.ID, err)
+				return fmt.Errorf("unable to delete cluster %q: %w", cluster.Name, err)
 			}
 
 			if res.ErrorCode != "200" {
-				return fmt.Errorf("Civo returned an error code %s when deleting cluster %q: %s", res.ErrorCode, cluster.ID, res.ErrorDetails)
+				return fmt.Errorf("Civo returned an error code %q when deleting cluster %q: %s", res.ErrorCode, cluster.Name, res.ErrorDetails)
 			}
 
-			outputwriter.WriteStdoutf("deleted cluster %q", cluster.ID)
+			outputwriter.WriteStdoutf("deleted cluster %q", cluster.Name)
 		} else {
-			c.logger.Printf("refusing to delete cluster %q: nuke is not enabled", cluster.ID)
+			c.logger.Warnf("refusing to delete cluster %q: nuke is not enabled", cluster.Name)
 		}
 
-		network, err := c.client.FindNetwork(cluster.Name)
+		network, err := c.client.FindNetwork(cluster.ID)
 		if err != nil {
-			return fmt.Errorf("unable to find network for cluster %q: %w", cluster.ID, err)
+			if errors.Is(err, civogo.ZeroMatchesError) {
+				c.logger.Warnf("no network found for cluster %q", cluster.Name)
+				continue
+			}
+
+			return fmt.Errorf("unable to find network for cluster %q: %w", cluster.Name, err)
+		}
+
+		if c.nameFilter != nil && !c.nameFilter.MatchString(network.Name) {
+			c.logger.Warnf("skipping network %q: name does not match filter", network.Name)
+			continue
 		}
 
 		if c.nuke {
-			c.logger.Printf("deleting network %q", network.ID)
+			c.logger.Infof("deleting network %q", network.Name)
 			res, err := c.client.DeleteNetwork(network.ID)
 			if err != nil {
-				return fmt.Errorf("unable to delete cluster network %q: %w", cluster.ID, err)
+				return fmt.Errorf("unable to delete cluster network %q: %w", cluster.Name, err)
 			}
 
 			if res.ErrorCode != "200" {
-				return fmt.Errorf("Civo returned an error code %s when deleting network %q: %s", res.ErrorCode, network.ID, res.ErrorDetails)
+				return fmt.Errorf("Civo returned an error code %q when deleting network %q: %s", res.ErrorCode, network.Name, res.ErrorDetails)
 			}
 
-			outputwriter.WriteStdoutf("deleted network %q", network.ID)
+			outputwriter.WriteStdoutf("deleted network %q", network.Name)
 		} else {
-			c.logger.Printf("refusing to delete network %q: nuke is not enabled", network.ID)
+			c.logger.Warnf("refusing to delete network %q: nuke is not enabled", network.Name)
 		}
 	}
